Fixed right side popping and partly leftside. Also added gravity

+++ RetroESP/src/actor.cpp
@@ -15,7 +15,7 @@ int Actor::getX()
 
 int Actor::getY()
 {
-    return y;
+    return static_cast<int>(y);
 }
